Implement ability to override the StatsdClient
This is part of the resolution for https://github.com/zio/zio-metrics-connectors/issues/67 and also, required for https://github.com/SneaksAndData/arcane-framework-scala/issues/117

This is an architectural change that can simplify the implementation of the DataDog metrics through the Unix Domain socket.

In the current implementation, the `datadogLayer` creates its own instance of the `StatsdClient`, instead of obtaining it
from the ZIO environment. This results in the DataDog layer being highly coupled with the StatsdClient implementation,
and it subsequently becomes impossible to use the `StatsdClient` with a different configuration.

This PR introduces the following changes:
- The class `StatsdConfig` has been renamed to `StatsdIpConfig` in order to differentiate it from the forthcoming configuration model.
- A new trait, `StatsdConfig`, has been introduced. This should be consumed by the StatsdClient instances, replacing the concrete class.
- The `StatsdClientIpConfig` trait has also been introduced, serving as an abstraction layer between the DataDog configuration and the Statsd configuration.
- The `datadogLayer` no longer creates its own instance of `StatsdClient`; instead, it obtains the instance provided by the ZIO environment.
- DatadogConfig now implements `StatsdClientIpConfig`. This adjustment is necessary for the `StatsdClient` configuration to use `DataDogConfig` in place of `StatsdIpConfig`.